弾が対象にぶつかったら消えるようにしました。by koko

+++ SouceTreeShooting/Assets/Scripts/BulletScripts/BulletCore.cs
@@ -4,12 +4,12 @@ using UnityEngine;
 
 public class BulletCore :MonoBehaviour, IUnitMove, IUnitAttack
 {
-    public int moveSpd { get; set; }
+    public float moveSpd { get; set; }
 
     public int atkpow { get; set; }
     public int atkLayer { get; set; }
 
-    public void Move(int spd)
+    public void Move(float spd)
     {
         if (atkLayer == 1)
         {
